improve dpkg and apt version parsing

+++ coverage.sh
@@ -3134,11 +3134,13 @@ rm /tmp/debian-chroot/etc/fstab
 rm /tmp/debian-chroot/etc/hostname
 rm /tmp/debian-chroot/etc/resolv.conf
 rm /tmp/debian-chroot/var/lib/dpkg/status
-rm /tmp/debian-chroot/var/lib/dpkg/available
 rm /tmp/debian-chroot/var/cache/apt/archives/lock
 rm /tmp/debian-chroot/var/lib/dpkg/lock
 rm /tmp/debian-chroot/var/lib/dpkg/lock-frontend
+if [ "$mode" != "chrootless" ] || dpkg --compare-versions "\$(dpkg --robot --version)" lt 1.20.0; then
+rm /tmp/debian-chroot/var/lib/dpkg/available
 rm /tmp/debian-chroot/var/lib/dpkg/cmethopt
+fi
 rm /tmp/debian-chroot/var/lib/apt/lists/lock
 ## delete merged usr symlinks
 #rm /tmp/debian-chroot/libx32
@@ -3202,13 +3204,15 @@ rm /tmp/debian-chroot/etc/fstab
 rm /tmp/debian-chroot/etc/hostname
 rm /tmp/debian-chroot/etc/resolv.conf
 rm /tmp/debian-chroot/var/lib/dpkg/status
-rm /tmp/debian-chroot/var/lib/dpkg/available
 rm /tmp/debian-chroot/var/cache/apt/archives/lock
 rm /tmp/debian-chroot/var/lib/dpkg/lock
 rm /tmp/debian-chroot/var/lib/dpkg/lock-frontend
-rm /tmp/debian-chroot/var/lib/dpkg/cmethopt
 rm /tmp/debian-chroot/var/lib/apt/lists/lock
 rm /tmp/debian-chroot/var/lib/apt/extended_states
+if [ "$mode" != "chrootless" ] || dpkg --compare-versions "\$(dpkg --robot --version)" lt 1.20.0; then
+rm /tmp/debian-chroot/var/lib/dpkg/available
+rm /tmp/debian-chroot/var/lib/dpkg/cmethopt
+fi
 ## delete merged usr symlinks
 #rm /tmp/debian-chroot/libx32
 #rm /tmp/debian-chroot/lib64
@@ -3322,13 +3326,15 @@ rm /tmp/debian-chroot/etc/fstab
 rm /tmp/debian-chroot/etc/hostname
 rm /tmp/debian-chroot/etc/resolv.conf
 rm /tmp/debian-chroot/var/lib/dpkg/status
-rm /tmp/debian-chroot/var/lib/dpkg/available
 rm /tmp/debian-chroot/var/cache/apt/archives/lock
 rm /tmp/debian-chroot/var/lib/dpkg/lock
 rm /tmp/debian-chroot/var/lib/dpkg/lock-frontend
-rm /tmp/debian-chroot/var/lib/dpkg/cmethopt
 rm /tmp/debian-chroot/var/lib/apt/lists/lock
 rm /tmp/debian-chroot/var/lib/apt/extended_states
+if dpkg --compare-versions "\$(dpkg --robot --version)" lt 1.20.0; then
+rm /tmp/debian-chroot/var/lib/dpkg/available
+rm /tmp/debian-chroot/var/lib/dpkg/cmethopt
+fi
 ## delete merged usr symlinks
 #rm /tmp/debian-chroot/libx32
 #rm /tmp/debian-chroot/lib64
@@ -3383,13 +3389,15 @@ rm /tmp/debian-chroot/etc/fstab
 rm /tmp/debian-chroot/etc/hostname
 rm /tmp/debian-chroot/etc/resolv.conf
 rm /tmp/debian-chroot/var/lib/dpkg/status
-rm /tmp/debian-chroot/var/lib/dpkg/available
 rm /tmp/debian-chroot/var/cache/apt/archives/lock
 rm /tmp/debian-chroot/var/lib/dpkg/lock
 rm /tmp/debian-chroot/var/lib/dpkg/lock-frontend
-rm /tmp/debian-chroot/var/lib/dpkg/cmethopt
 rm /tmp/debian-chroot/var/lib/apt/lists/lock
 rm /tmp/debian-chroot/var/lib/apt/extended_states
+if dpkg --compare-versions "\$(dpkg --robot --version)" lt 1.20.0; then
+rm /tmp/debian-chroot/var/lib/dpkg/available
+rm /tmp/debian-chroot/var/lib/dpkg/cmethopt
+fi
 ## delete merged usr symlinks
 #rm /tmp/debian-chroot/libx32
 #rm /tmp/debian-chroot/lib64
